This project was a great one.
For me, the general picture was always known, but this filled in so many details of what existence of/with AGI is like, enough to convince me that these ideas really are the way of the future. It has the simplest core that keeps on giving, like the stuff of legends. Now, it depends on other people to succeed, so it will fail.

I'll never be able to forget it, though.