`we.remote`-only now works as expected (not computing until a connection).
Fixed a lingering bug in human extension use (double-cancel in the same port not working).

Fixed an introduced bug in `observers.js` compilation (where double-injection can't happen because of `const FN=…`). And where we did not expect the stream to be a promise, causing video/audio streams to not get created.